+++ flutter/test/default_integrations_test.dart
@@ -4,11 +4,12 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
+import 'package:sentry_flutter/src/sentry_flutter_options.dart';
 
 import 'mocks.dart';
 
 void main() {
-  const MethodChannel _channel = MethodChannel('sentry_flutter');
+  const _channel = MethodChannel('sentry_flutter');
 
   TestWidgetsFlutterBinding.ensureInitialized();
 
@@ -26,7 +27,7 @@ void main() {
     // replace default error otherwise it fails on testing
     FlutterError.onError = (FlutterErrorDetails errorDetails) async {};
 
-    flutterErrorIntegration(fixture.hub, fixture.options);
+    FlutterErrorIntegration()(fixture.hub, fixture.options);
 
     final throwable = StateError('error');
     final details = FlutterErrorDetails(exception: throwable);
@@ -51,7 +52,7 @@ void main() {
     };
     FlutterError.onError = defaultError;
 
-    flutterErrorIntegration(fixture.hub, fixture.options);
+    FlutterErrorIntegration()(fixture.hub, fixture.options);
 
     final throwable = StateError('error');
     final details = FlutterErrorDetails(exception: throwable);
@@ -65,7 +66,7 @@ void main() {
   });
 
   test('FlutterError adds integration', () async {
-    flutterErrorIntegration(fixture.hub, fixture.options);
+    FlutterErrorIntegration()(fixture.hub, fixture.options);
 
     expect(true,
         fixture.options.sdk.integrations.contains('flutterErrorIntegration'));
@@ -74,7 +75,7 @@ void main() {
   test('nativeSdkIntegration adds integration', () async {
     _channel.setMockMethodCallHandler((MethodCall methodCall) async {});
 
-    final integration = nativeSdkIntegration(fixture.options, _channel);
+    final integration = NativeSdkIntegration(_channel);
 
     await integration(fixture.hub, fixture.options);
 
@@ -87,7 +88,7 @@ void main() {
       throw null;
     });
 
-    final integration = nativeSdkIntegration(fixture.options, _channel);
+    final integration = NativeSdkIntegration(_channel);
 
     await integration(fixture.hub, fixture.options);
 
@@ -98,7 +99,7 @@ void main() {
   test('loadContextsIntegration adds integration', () async {
     _channel.setMockMethodCallHandler((MethodCall methodCall) async {});
 
-    final integration = loadContextsIntegration(fixture.options, _channel);
+    final integration = LoadContextsIntegration(_channel);
 
     await integration(fixture.hub, fixture.options);
 
@@ -109,5 +110,5 @@ void main() {
 
 class Fixture {
   final hub = MockHub();
-  final options = SentryOptions();
+  final options = SentryFlutterOptions();
 }